refactor(models): format files
