+++ packages/models/src/index.ts
@@ -1,7 +1,7 @@
 export { coreConfigSchema, CoreConfigSchema } from './lib/core-config';
 export { uploadConfigSchema, UploadConfigSchema } from './lib/upload';
 export { pluginConfigSchema, PluginConfigSchema } from './lib/plugins';
-export { runnerOutputSchema, RunnerOutputSchema, runnerOutputAuditRefsPresentInPluginConfigs } from './lib/runner-output';
+export { runnerOutputSchema, RunnerOutputSchema, runnerOutputAuditRefsPresentInPluginConfigs, PluginsOutputSchema } from './lib/output';
 export { persistConfigSchema, PersistConfigSchema } from './lib/persist';
 export {
   categoryConfigSchema,
